# App/BackEnd/Conduit.API/src/Conduit.API/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : {
      "name" : "RealWorld",
      "url" : "https://realworld.io"
    },
    "description" : "Conduit API",
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "title" : "Conduit API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
  "paths" : {
...
    "/articles/feed" : {
      "get" : {
        "description" : "Get most recent articles from users you follow. Use query parameters to limit. Auth is required",
        "operationId" : "GetArticlesFeed",
        "parameters" : [ {
          "description" : "Limit number of articles returned (default is 20)",
          "explode" : true,
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "Offset/skip number of articles (default is 0)",
          "explode" : true,
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MultipleArticlesResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Get recent articles from users you follow",
        "tags" : [ "Articles" ]
      }
    },
    "/articles" : {
      "get" : {
        "description" : "Get most recent articles globally. Use query parameters to filter results. Auth is optional",
        "operationId" : "GetArticles",
        "parameters" : [ {
          "description" : "Filter by tag",
          "explode" : true,
          "in" : "query",
          "name" : "tag",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "Filter by author (username)",
          "explode" : true,
          "in" : "query",
          "name" : "author",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "Filter by favorites of a user (username)",
          "explode" : true,
          "in" : "query",
          "name" : "favorited",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "Limit number of articles returned (default is 20)",
          "explode" : true,
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 20,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "Offset/skip number of articles (default is 0)",
          "explode" : true,
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MultipleArticlesResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "summary" : "Get recent articles globally",
        "tags" : [ "Articles" ]
      },
      "post" : {
        "description" : "Create an article. Auth is required",
        "operationId" : "CreateArticle",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewArticleRequest"
              }
            }
          },
          "description" : "Article to create",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SingleArticleResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Create an article",
        "tags" : [ "Articles" ],
        "x-codegen-request-body-name" : "article"
      }
    },
    "/articles/{slug}" : {
      "delete" : {
        "description" : "Delete an article. Auth is required",
        "operationId" : "DeleteArticle",
        "parameters" : [ {
          "description" : "Slug of the article to delete",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : { },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Delete an article",
        "tags" : [ "Articles" ]
      },
      "get" : {
        "description" : "Get an article. Auth not required",
        "operationId" : "GetArticle",
        "parameters" : [ {
          "description" : "Slug of the article to get",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SingleArticleResponse"
                }
              }
            },
            "description" : "OK"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "summary" : "Get an article",
        "tags" : [ "Articles" ]
      },
      "put" : {
        "description" : "Update an article. Auth is required",
        "operationId" : "UpdateArticle",
        "parameters" : [ {
          "description" : "Slug of the article to update",
          "explode" : false,
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateArticleRequest"
              }
            }
          },
          "description" : "Article to update",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SingleArticleResponse"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : { },
            "description" : "Unauthorized"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "summary" : "Update an article",
        "tags" : [ "Articles" ],
        "x-codegen-request-body-name" : "article"
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Profile" : {
        "example" : {
          "image" : "image",
          "following" : true,
          "bio" : "bio",
          "username" : "username"
        },
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "following" : {
            "type" : "boolean"
          }
        },
        "required" : [ "bio", "following", "image", "username" ],
        "type" : "object"
      },
      "Article" : {
        "example" : {
          "tagList" : [ "tagList", "tagList" ],
          "createdAt" : "2000-01-23T04:56:07.000+00:00",
          "author" : {
            "image" : "image",
            "following" : true,
            "bio" : "bio",
            "username" : "username"
          },
          "description" : "description",
          "title" : "title",
          "body" : "body",
          "favoritesCount" : 0,
          "slug" : "slug",
          "updatedAt" : "2000-01-23T04:56:07.000+00:00",
          "favorited" : true
        },
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "tagList" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "favorited" : {
            "type" : "boolean"
          },
          "favoritesCount" : {
            "type" : "integer"
          },
          "author" : {
            "$ref" : "#/components/schemas/Profile"
          }
        },
        "required" : [ "author", "body", "createdAt", "description", "favorited", "favoritesCount", "slug", "tagList", "title", "updatedAt" ],
        "type" : "object"
      },
      "SingleArticleResponse" : {
        "example" : {
          "article" : {
            "tagList" : [ "tagList", "tagList" ],
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "description" : "description",
            "title" : "title",
            "body" : "body",
            "favoritesCount" : 0,
            "slug" : "slug",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00",
            "favorited" : true
          }
        },
        "properties" : {
          "article" : {
            "$ref" : "#/components/schemas/Article"
          }
        },
        "required" : [ "article" ],
        "type" : "object"
      },
      "MultipleArticlesResponse" : {
        "example" : {
          "articlesCount" : 6,
          "articles" : [ {
            "tagList" : [ "tagList", "tagList" ],
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "description" : "description",
            "title" : "title",
            "body" : "body",
            "favoritesCount" : 0,
            "slug" : "slug",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00",
            "favorited" : true
          }, {
            "tagList" : [ "tagList", "tagList" ],
            "createdAt" : "2000-01-23T04:56:07.000+00:00",
            "author" : {
              "image" : "image",
              "following" : true,
              "bio" : "bio",
              "username" : "username"
            },
            "description" : "description",
            "title" : "title",
            "body" : "body",
            "favoritesCount" : 0,
            "slug" : "slug",
            "updatedAt" : "2000-01-23T04:56:07.000+00:00",
            "favorited" : true
          } ]
        },
        "properties" : {
          "articles" : {
            "items" : {
              "$ref" : "#/components/schemas/Article"
            },
            "type" : "array"
          },
          "articlesCount" : {
            "type" : "integer"
          }
        },
        "required" : [ "articles", "articlesCount" ],
        "type" : "object"
      },
      "NewArticle" : {
        "example" : {
          "tagList" : [ "tagList", "tagList" ],
          "description" : "description",
          "title" : "title",
          "body" : "body"
        },
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "tagList" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "body", "description", "title" ],
        "type" : "object"
      },
      "NewArticleRequest" : {
        "example" : {
          "article" : {
            "tagList" : [ "tagList", "tagList" ],
            "description" : "description",
            "title" : "title",
            "body" : "body"
          }
        },
        "properties" : {
          "article" : {
            "$ref" : "#/components/schemas/NewArticle"
          }
        },
        "required" : [ "article" ],
        "type" : "object"
      },
      "UpdateArticle" : {
        "example" : {
          "description" : "description",
          "title" : "title",
          "body" : "body"
        },
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UpdateArticleRequest" : {
        "example" : {
          "article" : {
            "description" : "description",
            "title" : "title",
            "body" : "body"
          }
        },
        "properties" : {
          "article" : {
            "$ref" : "#/components/schemas/UpdateArticle"
          }
        },
        "required" : [ "article" ],
        "type" : "object"
      },
...
      "GenericErrorModel" : {
        "properties" : {
          "errors" : {
            "$ref" : "#/components/schemas/GenericErrorModel_errors"
          }
        },
        "required" : [ "errors" ],
        "type" : "object"
      },
      "GenericErrorModel_errors" : {
        "properties" : {
          "body" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "body" ],
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "Token" : {
        "description" : "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}